+++ src/inventorum.ebay/inventorum/ebay/apps/products/models.py
@@ -145,7 +145,7 @@ class EbayItemModelQuerySet(BaseQuerySet):
         :rtype EbayItemModelQuerySet
         """
         inv_id = EbayItemModel.clean_sku(sku)
-        return self.filter(product__inv_id=inv_id)
+        return self.filter(inv_product_id=inv_id)
 
     def by_account(self, account):
         """
@@ -251,12 +251,8 @@ class EbayItemModel(OrderableItemModel, BaseModel):
         return self.variations.exists()
 
     @property
-    def inv_product_id(self):
-        return self.product.inv_id
-
-    @property
     def sku(self):
-        return settings.EBAY_SKU_FORMAT.format(self.product.inv_id)
+        return settings.EBAY_SKU_FORMAT.format(self.inv_product_id)
 
 
 class EbayItemVariationModelQuerySet(BaseQuerySet):
@@ -267,8 +263,6 @@ class EbayItemVariationModelQuerySet(BaseQuerySet):
 
 
 class EbayItemVariationModel(OrderableItemModel, BaseModel):
-    inv_product_id = models.IntegerField(verbose_name="Inventorum product id")
-
     quantity = models.IntegerField(default=0)
     gross_price = MoneyField()
     tax_rate = TaxRateField()