+++ src/inventorum.ebay/inventorum/ebay/lib/core_api/tests/test_user_scoped_core_api_client.py
@@ -55,7 +55,7 @@ class IntegrationTestUserScopedCoreAPIClient(APITestCase):
     def test_get_products_without_ebay_meta(self):
         core_product = self.subject.get_product(StagingTestAccount.Products.SIMPLE_PRODUCT_ID)
 
-        self.assertEqual(core_product.id, StagingTestAccount.Products.SIMPLE_PRODUCT_ID)
+        self.assertEqual(core_product.inv_id, StagingTestAccount.Products.SIMPLE_PRODUCT_ID)
         self.assertEqual(core_product.name, "XtC Advanced 2 LTD")
         self.assertEqual(core_product.description, "GIANT XtC 27.5\" Advanced Carbon")
         self.assertEqual(core_product.gross_price, D("1999.99"))
@@ -71,7 +71,7 @@ class IntegrationTestUserScopedCoreAPIClient(APITestCase):
     def test_get_product_with_ebay_meta(self):
         core_product = self.subject.get_product(StagingTestAccount.Products.PRODUCT_WITH_EBAY_META_ID)
 
-        self.assertEqual(core_product.id, StagingTestAccount.Products.PRODUCT_WITH_EBAY_META_ID)
+        self.assertEqual(core_product.inv_id, StagingTestAccount.Products.PRODUCT_WITH_EBAY_META_ID)
         # non-channeled name: FastRoad CoMax
         self.assertEqual(core_product.name, "eBay: FastRoad CoMax")
         # non-channeled description: GIANT FastRoad CoMax Carbon